pay to bussiness added

+++ controller/orderController.js
@@ -2,90 +2,59 @@ const stripe = require("stripe")(process.env.STRIPE_KEY);
 const Order = require("../models/Order");
 const { User } = require("../models/User");
 const { Cart } = require("../models/Cart");
-const {customerPaymentTracker} = require("../models/admin/PaymentTracker/paymentIdTracker")
-const mongoose = require('mongoose');
+const random = require('alphanumeric')
+const {
+  customerPaymentTracker,
+  customerRedeemTracker,
+} = require("../models/admin/PaymentTracker/paymentIdTracker");
+const {
+  SaveMyCropTrasaction,
+} = require("../controller/customerCropTransaction");
+const mongoose = require("mongoose");
 // create-payment-intent
-// module.exports.paymentIntent = async (req, res) => { 
- 
-//   try {
-//     let {price} = req.body;
-    
-//     const token = req.headers?.authorization?.split(" ")?.[1];
-//     if(!price){
-//       return res.status(406).send({msg:"product price is required"})
-//     }
-//     if(!token){
-//       return res.status(406).send({msg:"you are not authorise"})
-//     }
-//      price = Number(price);
-//     const amount = price * 100;
-  
-//     // Create a PaymentIntent with the order amount and currency
-//     const paymentIntent = await stripe.paymentIntents.create({
-//       currency: "usd",
-//       amount: amount,
-//       payment_method_types: ['card'],
-//       description: 'Test Payment',
-//       statement_descriptor: 'TEST PAYMENT',
-//     });
-//     console.log(paymentIntent.client_secret, "payments")
-
-//   //  if(paymentIntent.client_secret)
-//   //  {
-//   //    const results=await User.updateOne({"token": token }, {$push: { auditTrail:`The payment is successfully paid` }});
-//   //  }
-//     res.status(200).send({
-//       clientSecret: paymentIntent.client_secret,
-//     });
-//   } 
-//   catch (error) {
-//     console.log(error);
-//   }
-// };
 
 module.exports.addOrder = async (req, res) => {
   try {
     const orderItems = req.body;
-    const croppoints=orderItems.points
-     const userid=orderItems.user;
+    const croppoints = orderItems.points;
+    const userid = orderItems.user;
 
     const newOrders = new Order(orderItems);
-    const userData=await User.findOne(
-      {
-        _id: userid       
-      });
+    const userData = await User.findOne({
+      _id: userid,
+    });
 
-    const points= userData.croppoints + croppoints;
+    const points = userData.croppoints + croppoints;
 
-    const result=await User.updateOne({_id: userid }, {$set: { croppoints:points }});
-    // const results=await User.updateOne({_id: userid }, {$push: { auditTrail:`The order is conformed` }});
-
+    const result = await User.updateOne(
+      { _id: userid },
+      { $set: { croppoints: points } }
+    );
     const order = await newOrders.save();
     res.status(200).send({
       success: true,
       message: "Order added successfully",
       order: order,
     });
-  } 
-  catch (error) {
+  } catch (error) {
     console.log(error);
   }
 };
 
-module.exports.paymentIntent= async(req, res)=> {
-  const {cart, user_id} = req.body;
+module.exports.paymentIntent = async (req, res) => {
+  const { cart, user_id } = req.body;
   // res.send(req.body);
   // console.log(req.body);
-  // return 
-  if (req.method === "POST") {    
+  // return
+  if (req.method === "POST") {
     try {
       const params = {
         submit_type: "pay",
         mode: "payment",
-        invoice_creation: {enabled: true},
-        payment_method_types: ["card", 'alipay'],
+        invoice_creation: { enabled: true },
+        payment_method_types: ["card", "alipay"],
         billing_address_collection: "auto",
-        line_items: cart.map((item) => {    
+        line_items: cart.map((item) => {
           return {
             price_data: {
               currency: "aud",
@@ -104,34 +73,32 @@ module.exports.paymentIntent= async(req, res)=> {
         }),
         success_url: `${req.headers.origin}/success`,
         cancel_url: `${req.headers.origin}/canceled`,
-        billing_address_collection: 'required',
+        billing_address_collection: "required",
         shipping_address_collection: {
-          allowed_countries: ['AU'],
+          allowed_countries: ["AU"],
         },
       };
-      
-      const session = await stripe.checkout.sessions.create(params);
-      if(session.id){   
-      await customerPaymentTracker.create({
-        paymentId:session.id,
-        status:"unpaid",
-        paymentMethod:session.payment_method_types,
-        paymentUrl:session.url,
-        cartDetails:{
-          id:req.body._id,
-          user_id:req.body.user_id,
-          cartItems:req.body.cart
-        }
-      })
 
-      // await Cart.updateMany(
-      //   { 'user_id': mongoose.Types.ObjectId(user_id) },
-      //   { $set: { 'cart.$[].purchaseStatus': 1 } }
-      // )
-      await Cart.deleteMany(
-        { 'user_id': mongoose.Types.ObjectId(user_id) }
-      )
-    }
+      const session = await stripe.checkout.sessions.create(params);
+      if (session.id) {
+        await customerPaymentTracker.create({
+          paymentId: session.id,
+          status: "unpaid",
+          paymentMethod: session.payment_method_types,
+          paymentUrl: session.url,
+          cartDetails: {
+            id: req.body._id,
+            user_id: req.body.user_id,
+            cartItems: req.body.cart,
+          },
+        });
+
+        // await Cart.updateMany(
+        //   { 'user_id': mongoose.Types.ObjectId(user_id) },
+        //   { $set: { 'cart.$[].purchaseStatus': 1 } }
+        // )
+        await Cart.deleteMany({ user_id: mongoose.Types.ObjectId(user_id) });
+      }
 
       res.status(200).json(session);
     } catch (err) {
@@ -141,6 +108,46 @@ module.exports.paymentIntent= async(req, res)=> {
     res.setHeader("Allow", "POST");
     res.status(405).end("Method Not Allowed");
   }
-}
+};
 
-//adding the croppoints is remaining
+module.exports.RedeemCrop = async (req, res) => {
+  const { cart, user_id, _id } = req.body;
+  let redeemCropPoints = 0;
+  cart?.map((x) => {
+    redeemCropPoints = redeemCropPoints + x.redeemCROPs;
+  });
+  console.log(redeemCropPoints);
+  let findUser = await User.findOne({ _id: user_id.$oid });
+
+  if (redeemCropPoints > findUser.croppoints) {
+    return res
+      .status(200)
+      .send({ msg: "sorry insoffient CROP in your account" });
+  }
+  let newCropPoint = findUser.croppoints - redeemCropPoints;
+  await User.findByIdAndUpdate(
+    { _id: findUser._id },
+    { $set: { croppoints: newCropPoint } }
+  );
+  let orderNumber = random(7)
+  await customerRedeemTracker.create({
+    number:orderNumber,
+    cartDetails: {
+      id: _id.$oid,
+      user_id: user_id.$oid,
+      cartItems: cart,
+    },
+  });
+  
+  SaveMyCropTrasaction(
+    1,
+    redeemCropPoints,
+    "debit",
+    "purchase product by redeem CROP",
+    orderNumber,
+    user_id.$oid
+  );
+  //
+  res.status(200).send({ msg: "CROP redemption success" });
+  return;
+};
